requirements.txt
python-telegram-bot==20.6
psycopg2-binary
Flask
requests
beautifulsoup4
python-dotenv
